# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the API for cuHacking 2020.",
    "version": "1.0.2",
    "title": "Magneton (cuHacking 2020 API)",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "api.cuhacking.com",
  "basePath": "/api",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/users": {
      "post": {
        "tags": [
          "Users"
        ],
        "summary": "Create a new user",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "User data",
            "required": true,
            "schema": {
              "type": "object",
              "required": [
                "username",
                "password"
              ],
              "properties": {
                "username": {
                  "type": "string"
                },
                "token": {
                  "type": "string"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User successfully created"
          },
          "500": {
            "description": "Failure in parsing the token or creating the user in Firestore"
          }
        },
        "security": [
          {
            "magneton_auth": [
              "admin"
            ]
          }
        ]
      },
      "get": {
        "tags": [
          "Users"
        ],
        "summary": "Get all users in the database",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Users successfully retrieved",
            "schema": {
              "type": "object",
              "properties": {
                "operation": {
                  "type": "string"
                },
                "status": {
                  "type": "string"
                },
                "items": {
                  "type": "integer"
                },
                "data": {
                  "type": "object",
                  "properties": {
                    "username": {
                      "type": "string"
                    },
                    "role": {
                      "type": "string"
                    },
                    "uid": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error retrieving from Firestore"
          }
        },
        "security": [
          {
            "magneton_auth": [
              "admin"
            ]
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "magneton_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://cuhacking.com/users/signin",
      "flow": "implicit",
      "scopes": {
        "user": "Access own data",
        "admin": "Admin access"
      }
    }
  },
...
}